+++ generators/app/templates/webpack.production.config.js
@@ -11,7 +11,7 @@ module.exports = {
   context: __dirname,
   entry: './app/src/scripts/client.<%= jsExt %>',
   output: {
-    path: __dirname + '/app/assets/dist',
+    path: `${__dirname}/app/static/dist`,
     filename: 'bundle.js'
   },
   resolve: {